backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.17
numpy==1.26.4
opencv-python-headless==4.10.0.84
Pillow==10.4.0
mediapipe==0.10.14
